+++ NEWS.md
@@ -1,3 +1,7 @@
+# Rgemini 0.3.1
+
+* Allows the user to specify the number of digits to round percentages to in cell suppression of categorical variables, without forcing rounding to integer.
+
 # Rgemini 0.3.0
 
 * Exports `n_rbc_transfusion` function.